Version: 2011-01-27a ksh93u
Sourced from ast-open-history version 2011-01-18 and:
7b301fa2ea22f2698b8858c4fbd9712741d5b2806e2c32130e1de712d0395354  INIT.2011-01-27a.tgz
35c720ed641745009d04f33052e378251b41e87a8d338911b1b3d4d7eabd74ac  ast-ksh.2011-01-27a.tgz

+++ src/cmd/ksh93/sh/init.c
@@ -54,7 +54,7 @@
 #if !_typ_wctrans_t
 #undef	wctrans_t
 #define wctrans_t	sh_wctrans_t
-typedef int wctrans_t;
+typedef long wctrans_t;
 #endif
 #if !_lib_wctrans
 #undef	wctrans
